Minor patch to last commit

+++ zephyrapp/game/src/cat_sprite.c
@@ -693,9 +693,9 @@ void CAT_depthberry()
 
 void CAT_triberry
 (
-	int xa, int ya, uint16_t za,
-	int xb, int yb, uint16_t zb,
-	int xc, int yc, uint16_t zc,
+	int xa, int ya, float za,
+	int xb, int yb, float zb,
+	int xc, int yc, float zc,
 	uint16_t c
 )
 {
